chore: rector 2/2: application de rector au code du projet

+++ src/Debug/Toolbar/Collectors/FilesCollector.php
@@ -38,7 +38,7 @@ class FilesCollector extends BaseCollector
      */
     public function getTitleDetails(): string
     {
-        return '( ' . (int) count(get_included_files()) . ' )';
+        return '( ' . count(get_included_files()) . ' )';
     }
 
     /**
